+++ packages/core/src/utils/event-handlers-factory.js
